# One Step Simplifier: several optimizations