--- a/key.core/src/main/java/de/uka/ilkd/key/logic/PosInOccurrence.java
+++ b/key.core/src/main/java/de/uka/ilkd/key/logic/PosInOccurrence.java
@@ -187,6 +187,10 @@ public final class PosInOccurrence {
         return subTermCache;
     }
 
+    public boolean subTermExists() {
+        return posInTerm.existsSubTerm(sequentFormula.formula());
+    }
+
     /**
      * Ascend to the top node of the formula this object points to
      */