--- a/key.core/src/main/java/de/uka/ilkd/key/logic/PosInOccurrence.java
+++ b/key.core/src/main/java/de/uka/ilkd/key/logic/PosInOccurrence.java
@@ -122,6 +122,7 @@ public final class PosInOccurrence {
         return posInTerm.getIndex();
     }
 
+    @Override
     public int hashCode() {
         return hashCode;
     }
@@ -187,6 +188,9 @@ public final class PosInOccurrence {
         return subTermCache;
     }
 
+    /**
+     * @return whether this PiO is a valid reference
+     */
     public boolean subTermExists() {
         return posInTerm.existsSubTerm(sequentFormula.formula());
     }