--- a/key.core/src/main/java/de/uka/ilkd/key/macros/SMTPreparationMacro.java
+++ b/key.core/src/main/java/de/uka/ilkd/key/macros/SMTPreparationMacro.java
@@ -51,6 +51,11 @@ public class SMTPreparationMacro extends SequentialProofMacro {
         return "Auto Pilot";
     }
 
+    @Override
+    public String getScriptCommandName() {
+        return "smt-prepare";
+    }
+
     /**
      * Gets the description of this macro.
      * <p>