--- a/key.core/src/main/java/de/uka/ilkd/key/proof/Statistics.java
+++ b/key.core/src/main/java/de/uka/ilkd/key/proof/Statistics.java
@@ -12,7 +12,6 @@ import de.uka.ilkd.key.informationflow.proof.InfFlowProof;
 import de.uka.ilkd.key.informationflow.proof.SideProofStatistics;
 import de.uka.ilkd.key.proof.reference.ClosedBy;
 import de.uka.ilkd.key.rule.*;
-import de.uka.ilkd.key.rule.OneStepSimplifier.Protocol;
 import de.uka.ilkd.key.rule.merge.MergeRuleBuiltInRuleApp;
 import de.uka.ilkd.key.smt.SMTRuleApp;
 import de.uka.ilkd.key.util.EnhancedStringBuffer;
@@ -328,10 +327,10 @@ public class Statistics {
          */
         private int tmpOssCaptured(final RuleApp ruleApp) {
             int tmpOssCaptured = 0;
-            final Protocol protocol =
-                ((de.uka.ilkd.key.rule.OneStepSimplifierRuleApp) ruleApp).getProtocol();
-            if (protocol != null) {
-                tmpOssCaptured = protocol.size() - 1;
+            int protocol =
+                ((de.uka.ilkd.key.rule.OneStepSimplifierRuleApp) ruleApp).getProtocolLength();
+            if (protocol != 0) {
+                tmpOssCaptured = protocol - 1;
             }
             return tmpOssCaptured;
         }