--- a/key.core/src/main/resources/de/uka/ilkd/key/proof/rules/genericRules.key
+++ b/key.core/src/main/resources/de/uka/ilkd/key/proof/rules/genericRules.key
@@ -219,7 +219,7 @@
         \find((C)castedTerm)
         \replacewith(castedTerm)
         \displayname "castDel"
-        \heuristics(simplify, cast_deletion)
+        \heuristics(simplify, cast_deletion, oss)
     };
 
     castDel2 {
@@ -321,7 +321,7 @@
         \varcond(\strict \sub(Null, G), \strict \sub(Null, H), \disjointModuloNull(G, H))
 
         \replacewith(x = null & y = null)
-        \heuristics(simplify)
+        \heuristics(simplify, oss)
     };
 
     sortsDisjoint1 {