--- a/key.core/src/main/resources/de/uka/ilkd/key/proof/rules/ifThenElseRules.key
+++ b/key.core/src/main/resources/de/uka/ilkd/key/proof/rules/ifThenElseRules.key
@@ -60,7 +60,7 @@
     ifthenelse_negated {
         \find(\if(!phi) \then(then) \else(else))
         \replacewith(\if(phi) \then(else) \else(then))
-        \heuristics(simplify)
+        \heuristics(simplify, oss)
     };
 
     ifthenelse_negated_for {