--- a/key.core/src/main/resources/de/uka/ilkd/key/proof/rules/ruleSetsDeclarations.key
+++ b/key.core/src/main/resources/de/uka/ilkd/key/proof/rules/ruleSetsDeclarations.key
@@ -574,4 +574,7 @@ The optimized rules pull out select terms before they evaluate them. Because oth
 
     // double
     executeDoubleAssignment;
+
+    // One Step Simplifier
+    oss;
 }