… key.core/src/main/resources/de/uka/ilkd/key/proof/rules/heapRules.key
… key.core/src/main/resources/de/uka/ilkd/key/proof/rules/locSetsRules.key
… key.core/src/test/resources/de/uka/ilkd/key/nparser/taclets.old.txt
… key.core/tacletProofs/locSet/Taclet_disjointArrayRangeAllFields1.proof
… key.core/tacletProofs/locSet/Taclet_disjointArrayRangeAllFields2.proof
… key.core/tacletProofs/locSet/Taclet_disjointArrayRanges.proof
… key.core/tacletProofs/seqPerm/Taclet_seqPermFromSwap.proof
… key.core/tacletProofs/seqRules/Taclet_getOfArray2seq.proof
… key.core/tacletProofs/seqRules/Taclet_indexOfSeqSingleton.proof
… key.core/tacletProofs/seqRules/Taclet_seqDefOfSeq.proof
… key.core/tacletProofs/seqRules/Taclet_seqDef_split.proof
… key.core/tacletProofs/seqRules/Taclet_seqDef_split_in_three.proof
… key.core/tacletProofs/seqRules/Taclet_seqOutsideValue.proof
… key.ui/examples/heap/permissions/threads/Plotter_joinTransfer_accessible.proof
… key.ui/examples/heap/permissions/threads/Plotter_startTransfer_accessible.proof
… key.ui/examples/heap/permutedSum/perm.proof
… key.ui/examples/heap/verifyThis15_2_ParallelGcd/parallelGcd.proof
… key.ui/examples/heap/verifyThis15_3_DLL/doUndo.proof
… keyext.slicing/src/test/java/org/key_project/slicing/EndToEndTests.java