This PR fixes #3248 and supersedes #3256 and #3257 by optimizing the One Step Simplifier in several ways. See the commits in this branch for details.

Additionally, the following rules are added to a new `oss` ruleset included in the OSS:

- selectOfStore, selectOfCreate, selectOfAnon, selectOfMemset, selectCreatedOfAnon, dismissNonSelectedField
- elementOfUnion, elementOfArrayRange
- sortsDisjointModuloNull
- ifthenelse_negated
- castDel

This means that the One Step Simplifier can now simplify many heap terms in a single step. Previously this required many more proof nodes. For the file linked below, this additional ruleset is required (see https://github.com/KeYProject/key/issues/3248#issuecomment-1704775679).

Benchmark (Finish Symbolic Execution) on [ShortestPath.java](https://gist.github.com/FliegendeWurst/9be15e20e3bfd0722fcdd7f5c4afebcd).

| Description                   | Nodes | Rule apps | Time |
:------------------------------:|:-----:|:---------:|:-----:
Baseline                        | 2341  | 10,289    | 18.3s
hashCode: op() and depth()      | 2341  | 10,289    | 18.0s
inner loop on current subterm   | 2341  | 16,986    | 18.1s
replaceKnown: skip unchanged    | 2341  | 16,986    | 18.1s
replaceKnown on demand          | 2341  | 16,986    | 17.2s
optimized cycle check           | 2341  | 16,986    | 17.2s
with new rules                  | 2229  | 131,671   | 33.5s
with new rules (no cycle check) | 2229  | 131,671   | 32.7s

cc @WolframPfeifer @mattulbrich